combine the whole tests in one file
