--- Tests/test_read_root.py
@@ -1,12 +1,0 @@
-import pytest
-
-from tests.test import client 
-
-
-
-# Test the root endpoint
-def test_read_root():
-    response = client.get("/")
-    assert response.status_code == 200
-    assert response.json() == {"message": "Welcome to the Emotion detection API"}
-
